% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{aggregateContributorData}
\alias{aggregateContributorData}
\title{Gets oldest commit date and handle for provided contributor paths for a Jekyll/Github Pages repository.}
\usage{
aggregateContributorData(contributorData)
}
\arguments{
\item{contributorData}{takes the output of `getContributorDataJekyll()`. See docs for description of output.}
}
\description{
Gets oldest commit date and handle for provided contributor paths for a Jekyll/Github Pages repository.
}
