% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcCommitNum}
\alias{calcCommitNum}
\title{Calculates the number of commits/month to a given repository over a specified date range.}
\usage{
calcCommitNum(commitObjList)
}
\arguments{
\item{commitObjList}{The result from `getCommits()` or `gh("GET /repos/:owner/:repo/commits, ... )`. A list where each element contains all info relevant to a single commit.}
}
\value{
a tibble that shows the number of commits to a repository by month. Tibble has two columns: `month`, `numCommits`.
}
\description{
Calculates the number of commits/month to a given repository over a specified date range.
}
