% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcContributorNum}
\alias{calcContributorNum}
\title{This function calculates the number of new contributors and total contributors}
\usage{
calcContributorNum(owner, repo, dateRange)
}
\arguments{
\item{owner}{The owner of the repository that the blog lives in.}

\item{repo}{The repository name.}

\item{dateRange}{A vector of two dates.}
}
\value{

}
\description{
This function calculates the number of new contributors and total contributors
}
