% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcContributorStats}
\alias{calcContributorStats}
\title{This function calculates statistics such as the total number of contributors, number of new contributors, and their names for a given month.}
\usage{
calcContributorStats(contributorDateDf, month, year)
}
\arguments{
\item{contributorDateDf}{The owner of the repository to pull file paths from. Defaults to "FredHutch".}

\item{year}{}
}
\value{

}
\description{
This function calculates statistics such as the total number of contributors, number of new contributors, and their names for a given month.
}
