% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcPostNum}
\alias{calcPostNum}
\title{Calculate the number of posts per month in a Jekyll/GitHub pages repository.}
\usage{
calcPostNum(owner, repo, dateRange, by = "postName")
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}

\item{dateRange}{A vector of two dates.}

\item{by}{"postNames" is the only option currently. Until I can determine a better way to calculate posts.}
}
\value{
a dataframe of number of posts per month
}
\description{
Calculate the number of posts per month in a Jekyll/GitHub pages repository.
}
