% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcPostNumJekyll}
\alias{calcPostNumJekyll}
\title{Calculate the number of posts per month in a Jekyll GitHub pages repository.}
\usage{
calcPostNumJekyll(posts, dateRange)
}
\arguments{
\item{posts}{a vector of post file names.}

\item{dateRange}{A vector of two dates in yyyy-mm-dd format. Can be strings or date objects. Order doesn't matter, the max and min will be used.}
}
\value{
a dataframe of number of posts per month
}
\description{
Calculate the number of posts per month in a Jekyll GitHub pages repository.
}
