% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{calcTotalCommits}
\alias{calcTotalCommits}
\title{Parses an object returned by `getCommitActivity` and calculates the total number of commits that occured during the time frame encapsulated in the object.}
\usage{
calcTotalCommits(commitObj)
}
\arguments{
\item{commitObj}{Object returned from an `gh` API pull of commit activity.}
}
\value{
The total number of commits
}
\description{
Parses an object returned by `getCommitActivity` and calculates the total number of commits that occured during the time frame encapsulated in the object.
}
