% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{checkCacheExists}
\alias{checkCacheExists}
\title{Checks whether or not the internal data cache (`R/sysdata.rda`) exists.}
\usage{
checkCacheExists()
}
\value{
TRUE/FALSE. TRUE, data cache exists. FALSE, data cache does not exist.
}
\description{
Checks whether or not the internal data cache (`R/sysdata.rda`) exists.
}
