% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{completeMonths}
\alias{completeMonths}
\title{Fill in missing months in a vector or dataframe column. If a dataframe is supplied the extra columns will be filled in with NA. Dates should be a date object.}
\usage{
completeMonths(data, dateCol = NULL)
}
\arguments{
\item{data}{A vector or dataframe of dates in yyyy-mm-dd format. Filled in dates will be for the first of the month.}

\item{dateCol}{if a dataframe is supplied, what is the column name of the dates variable. Defaults to "month".}
}
\description{
Fill in missing months in a vector or dataframe column. If a dataframe is supplied the extra columns will be filled in with NA. Dates should be a date object.
}
