% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{contributorMetrics}
\alias{contributorMetrics}
\title{This function calculates the number of new contributors and total contributors by month}
\usage{
contributorMetrics(owner, repo, dateRange, useCache = FALSE)
}
\arguments{
\item{owner}{The owner of the repository that the blog lives in.}

\item{repo}{The repository name.}

\item{dateRange}{A vector of two dates in yyyy-mm-dd format. Can be strings or date objects. Order doesn't matter, the max and min will be used.}

\item{useCache}{TRUE/FALSE. When TRUE, use cached data. When FALSE, pull all data. Using cached data will make the function run faster.}
}
\value{

}
\description{
This function calculates the number of new contributors and total contributors by month
}
