% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{createCache}
\alias{createCache}
\title{Create a cache of data. This allows functions to run more quickly. Data is saved internally to the package as `R/sysdata.rda`. Currently, data cache only contains contributor data.}
\usage{
createCache(owner, repo, dateRange, overwrite = FALSE)
}
\arguments{
\item{owner}{The owner of the repository that the blog lives in.}

\item{repo}{The repository name.}

\item{dateRange}{A vector of two dates in yyyy-mm-dd format. Can be strings or date objects. Order doesn't matter.}

\item{overwrite}{TRUE/FALSE. If TRUE, overwrites current data cache. If FALSE, will not overwrite current cache. If there is no existing cache this argument does nothing.}
}
\value{
A dataframe
}
\description{
Create a cache of data. This allows functions to run more quickly. Data is saved internally to the package as `R/sysdata.rda`. Currently, data cache only contains contributor data.
}
