% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{createContributorCache}
\alias{createContributorCache}
\title{Create a cache of contribuor data. This allows other functions to run more quickly. Data is saved internally as `R/sysdata.rda`. This function pulls all data from the last two years.}
\usage{
createContributorCache(owner, repo, dateRange, overwrite)
}
\arguments{
\item{owner}{The owner of the repository that the blog lives in.}

\item{repo}{The repository name.}

\item{dateRange}{A vector of two dates in yyyy-mm-dd format. Can be strings or date objects. Order doesn't matter.}

\item{overwrite}{TRUE/FALSE. If TRUE, overwrites current data cache. If FALSE, will not overwrite current cache. If there is no existing cache this argument does nothing.}
}
\value{
A dataframe
}
\description{
Create a cache of contribuor data. This allows other functions to run more quickly. Data is saved internally as `R/sysdata.rda`. This function pulls all data from the last two years.
}
