% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{getAccountInfo}
\alias{getAccountInfo}
\title{Given a Google Analytics `webPropertyName` this function will pull the corresponding GoogleAnalytics account information}
\usage{
getAccountInfo(webPropertyName, onlyViewId = TRUE)
}
\arguments{
\item{webPropertyName}{a Google Analytics `webPropertyName`. You can find out the webproperty name using `ga_account_list()`.}

\item{year}{specify year as `yyyy`. Defaults to current year.}
}
\value{
a dataframe of metrics for the month and year specified.
}
\description{
Given a Google Analytics `webPropertyName` this function will pull the corresponding GoogleAnalytics account information
}
