% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogMetrics.R
\name{getBlogStatistics}
\alias{getBlogStatistics}
\title{Pulls blog statistics for blogs using the GitHub Pages directory structure from both GitHub and Google Analytics.}
\usage{
getBlogStatistics(webPropertyName, owner, repo, dateRange, useCache)
}
\arguments{
\item{webPropertyName}{A variable from GoogleAnalytics. You can find out the webproperty name using `ga_account_list()` form the `GoogleAnalyticsR` package.}

\item{owner}{The owner of the relevant GitHub repo.}

\item{repo}{The repository name of the relevant GitHub repo.}

\item{dateRange}{A vector of two dates.}
}
\value{
a dataframe of metrics for the month and year specified.
}
\description{
Pulls blog statistics for blogs using the GitHub Pages directory structure from both GitHub and Google Analytics.
}
