% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getCommitActivity}
\alias{getCommitActivity}
\title{Get an object of commit activity from a specified repository}
\usage{
getCommitActivity(owner = "FredHutch", repo = "coop")
}
\arguments{
\item{owner}{string of the repository owner. Default "FredHutch".}

\item{repo}{string of the repository name. Default "coop".}
}
\description{
Get an object of commit activity from a specified repository
}
