% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getCommitObj}
\alias{getCommitObj}
\title{Lists commits for specified repository and dateRange. Set to return the max amount per page (100).}
\usage{
getCommitObj(owner, repo, dateRange, pageNum)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}

\item{dateRange}{A vector of two dates.}

\item{pageNum}{Page number of the results to return.}
}
\value{
a date object of the date that the specified file was first commited to the repository.
}
\description{
Lists commits for specified repository and dateRange. Set to return the max amount per page (100).
}
