% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getCommits}
\alias{getCommits}
\title{Lists commits for a specified repository over a date range.}
\usage{
getCommits(owner, repo, start, end)
}
\arguments{
\item{owner}{The GitHub repository owner.}

\item{repo}{The GitHub repository name.}

\item{start}{Beginning of date range in yyyy-mm-dd format. Can be strings or date objects.}

\item{end}{End of date range in yyyy-mm-dd format. Can be strings or date objects.}
}
\value{
A list where each element contains information about a single commit including author, date, and commit message.
}
\description{
Lists commits for a specified repository over a date range.
}
