% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getContributorData}
\alias{getContributorData}
\title{Gets oldest commit date and handle for provided contributor path for a Jekyll/Github Pages repository. Used to pull data into the DATA.R script.}
\usage{
getContributorData(contributorPath, owner = "FredHutch", repo = "coop")
}
\arguments{
\item{contributorPath}{A path to a contributor file or directory containing contributor files.}

\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}
}
\description{
Gets oldest commit date and handle for provided contributor path for a Jekyll/Github Pages repository. Used to pull data into the DATA.R script.
}
