% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getContributorDataJekyll}
\alias{getContributorDataJekyll}
\title{Gets oldest commit date and handle for provided contributor paths for a Jekyll/Github Pages repository.}
\usage{
getContributorDataJekyll(owner, repo, useCache)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}

\item{useCache}{If TRUE, only returns new, uncached contributors. If FALSE, returns all contributors.}
}
\value{
A tibble with three columns: `path` contains the relative path to the contributor .md file, `commitDate` contains the earliest commit date for each contributor file, and `handle` has the contributors handle (the filename without .md on the end)
}
\description{
Gets oldest commit date and handle for provided contributor paths for a Jekyll/Github Pages repository.
}
