% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getContributorPathsJekyll}
\alias{getContributorPathsJekyll}
\title{Pulls contributor paths from a static Jekyll blog hosted on GitHub pages.}
\usage{
getContributorPathsJekyll(owner, repo, onlyUncached)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}

\item{useCache}{If TRUE, only returns new, uncached contributors. If FALSE, returns all contributors.}
}
\description{
Pulls contributor paths from a static Jekyll blog hosted on GitHub pages.
}
