% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getDirContents}
\alias{getDirContents}
\title{Get the names of files in a specified directory in a github repo}
\usage{
getDirContents(owner, repo, dir, fullPath = TRUE)
}
\arguments{
\item{owner}{the GitHub repository owner}

\item{repo}{The GitHub repository name}

\item{dir}{The directory name or path to directory}

\item{fullPath}{TRUE returns the full filepath, FALSE returns filename}
}
\value{
a vector of files located in the specified directory.
}
\description{
Uses the `gh` package to pull the file names from a specified directory.
}
