% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getFileNames}
\alias{getFileNames}
\title{Get the names of files in a specified directory in a github pages repo}
\usage{
getFileNames(owner = "FredHutch", repo = "coop", path = NULL)
}
\arguments{
\item{owner}{github pages repo owner. Default = "FredHutch".}

\item{repo}{github pages repo name. Default = "coop".}

\item{path}{github pages path (including "_") of directory.}
}
\value{
a vector of file names from the specified directory in `path` argument.
}
\description{
Uses the `gh` package to pull the file names of files from the specified directory.
}
