% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getGithubStatistics}
\alias{getGithubStatistics}
\title{Pull GitHub Repository Statistics data from an account.}
\usage{
getGithubStatistics()
}
\value{
a dataframe of variables
}
\description{
Uses the `gh` package to pull specified data from the GitHubRepository Statistics API
}
