% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{getPageViewsByPath}
\alias{getPageViewsByPath}
\title{Using the Google analytics API pulls page views by page path for the date range specified.}
\usage{
getPageViewsByPath(viewId, dateRange, onlyPosts = TRUE)
}
\arguments{
\item{viewId}{A Google Analytics `viewId`. You can find out the webproperty name using `getAccountInfo()` or `ga_account_list()`.}

\item{dateRange}{A vector of two dates.}

\item{onlyPosts}{A binary parameter. If TRUE only returns page path results for posts.}
}
\value{
a dataframe of metrics for the month and year specified.
}
\description{
Using the Google analytics API pulls page views by page path for the date range specified.
}
