% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getPaths}
\alias{getPaths}
\title{Given the repository owner and name and a specific path within that repo this function will return a vector of filepaths within that directory.}
\usage{
getPaths(owner = "FredHutch", repo = "coop", path = "_contributors")
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from. Defaults to "FredHutch".}

\item{repo}{The name of the repository to pull file paths from. Defaults to "coop".}

\item{path}{The directory name to pull file paths from. Defaults to "_contributors".}
}
\value{
a vector of file paths.
}
\description{
Given the repository owner and name and a specific path within that repo this function will return a vector of filepaths within that directory.
}
