% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getPostNames}
\alias{getPostNames}
\title{Get the names of posts from your github pages _posts repo}
\usage{
getPostNames(owner = "FredHutch", repo = "coop")
}
\arguments{
\item{owner}{github pages repo owner. Default = "FredHutch".}

\item{repo}{github pages repo name. Default = "coop".}
}
\value{
a vector of postNames from the `_posts` directory
}
\description{
Uses the `gh` package to pull the names of posts from the `_posts` directory
}
