% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{getReportData}
\alias{getReportData}
\title{Compares posts from this month to last month using the Google Analytics API}
\usage{
getReportData()
}
\value{
a dataframe of variables
}
\description{
Uses the `googleAnalyticsR` package to pull specified data from the GitHubRepository Statistics API
}
