% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getUncachedContributorPath}
\alias{getUncachedContributorPath}
\title{Compares current contributor files in repository to known cache of contributors.}
\usage{
getUncachedContributorPath(
  owner = "FredHutch",
  repo = "coop",
  knownContributors
)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}
}
\description{
Compares current contributor files in repository to known cache of contributors.
}
