% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{githubMetrics}
\alias{githubMetrics}
\title{Pull repository statistics for specified GitHub repository. Specifically repositories set up to host a Jekyll/GitHub Pages site.}
\usage{
githubMetrics(owner, repo, dateRange, useCache)
}
\arguments{
\item{owner}{The repository name of the relevant GitHub repo.}

\item{repo}{The owner of the relevant GitHub repo.}

\item{dateRange}{A vector of two dates.}

\item{useCache}{TRUE/FALSE.}
}
\value{
a dataframe of metrics for the date range specified.
}
\description{
Pull repository statistics for specified GitHub repository. Specifically repositories set up to host a Jekyll/GitHub Pages site.
}
