% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{listFilesandEarliestCommitDate}
\alias{listFilesandEarliestCommitDate}
\title{This function pulls the earliest commit date for each file in a specified path.}
\usage{
listFilesandEarliestCommitDate(
  owner = "FredHutch",
  repo = "coop",
  path = "_contributors",
  ordered = TRUE
)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from. Defaults to "FredHutch".}

\item{repo}{The name of the repository to pull file paths from. Defaults to "coop".}

\item{path}{The directory name to pull file paths from. Defaults to "_contributors".}

\item{ordered}{A binary parameter. If set to TRUE the dataframe returned will be ordered by newest to oldest file.}
}
\value{
a date object of the date that the specified file was first commited to the repository.
}
\description{
This function pulls the earliest commit date for each file in a specified path.
}
