% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{monthYear2DateRange}
\alias{monthYear2DateRange}
\title{Takes in a month and year and outputs the date range of that month as a list.}
\usage{
monthYear2DateRange(month, year)
}
\arguments{
\item{month}{specify month as `mm` or `m`. Defaults to current month.}

\item{year}{specify year as `yyyy`. Defaults to current year.}
}
\value{
a list of the date range for the specified month/year.
}
\description{
Takes in a month and year and outputs the date range of that month as a list.
}
