% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{newFilesWithinDateRange}
\alias{newFilesWithinDateRange}
\title{When given the output of `listFilesAndEarliestCommitDate()` and a date range this function returns the files that were commited within the specified date range.}
\usage{
newFilesWithinDateRange(fileDateDf, first, last)
}
\arguments{
\item{fileDateDf}{The owner of the repository to pull file paths from. Defaults to "FredHutch".}

\item{first}{The date object of the beginning of the date range.}

\item{last}{The date object of the end of the date range.}
}
\value{
vector of files that were first commited to the repository within the specified date range.
}
\description{
When given the output of `listFilesAndEarliestCommitDate()` and a date range this function returns the files that were commited within the specified date range.
}
