% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{path2Contributor}
\alias{path2Contributor}
\title{Given paths from the `_contributors` directory this function will return just the contributor ID.}
\usage{
path2Contributor(contributorPaths)
}
\arguments{
\item{contributorPaths}{A vector of paths from the `_contributors` directory.}
}
\value{
A vector of contributor IDs from the `_contributors` directory.
}
\description{
Given paths from the `_contributors` directory this function will return just the contributor ID.
}
