% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{path2OldestCommitDate}
\alias{path2OldestCommitDate}
\title{This function pulls the the oldest commit date for the specified path.}
\usage{
path2OldestCommitDate(owner = owner, repo = repo, path = path)
}
\arguments{
\item{owner}{The owner of the repository to pull file paths from.}

\item{repo}{The name of the repository to pull file paths from.}

\item{path}{The directory name to pull file paths from.}
}
\value{
a date object of the date that the specified file was first commited to the repository.
}
\description{
This function pulls the the oldest commit date for the specified path.
}
