% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{pathToContributor}
\alias{pathToContributor}
\title{Given a filepath from the `_contributors` directory this function will return just the contributor ID.}
\usage{
pathToContributor(contributorPaths)
}
\arguments{
\item{contributorPaths}{A vector of paths from the `_contributors` directory.}
}
\value{
A vector of contributor IDs from the `_contributors` directory.
}
\description{
Given a filepath from the `_contributors` directory this function will return just the contributor ID.
}
