% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{postsThisMonth}
\alias{postsThisMonth}
\title{Get the post names of post from the specified month and year (if left unspecified the current month/year is assumed)}
\usage{
postsThisMonth(fileNames, month = NULL, year = NULL)
}
\arguments{
\item{fileNames}{vector of file names in `yyyy-mm-dd-name.md` format}

\item{month}{month to get postNames for in `m` format}

\item{year}{year to get postNames for in `yyyy` format}
}
\value{
a df of posts names from this months with other variables
}
\description{
Gets postNames from month (`m`) and year (`yyyy`) specified
}
