% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{pullGithub}
\alias{pullGithub}
\title{Pull repository statistics for specified GitHub repository. Specifically repositories set up to host a Jekyll/GitHub Pages site.}
\usage{
pullGithub(owner = "FredHutch", repo = "coop", dateRange)
}
\arguments{
\item{owner}{The repository name of the relevant GitHub repo.}

\item{repo}{The owner of the relevant GitHub repo.}

\item{dateRange}{A vector of two dates.}
}
\value{
a dataframe of metrics for the date range specified.
}
\description{
Pull repository statistics for specified GitHub repository. Specifically repositories set up to host a Jekyll/GitHub Pages site.
}
