% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{pullGoogleAnalytics}
\alias{pullGoogleAnalytics}
\title{API pull from both Google Analytics.}
\usage{
pullGoogleAnalytics(webPropertyName, dateRange)
}
\arguments{
\item{webPropertyName}{A variable from GoogleAnalytics. You can find out the webproperty name using `ga_account_list()`}

\item{dateRange}{A vector of two dates.}
}
\value{
a dataframe of monthly metrics from the date range specified.
}
\description{
API pull from both Google Analytics.
}
