% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{subsetCommitActivityByMonth}
\alias{subsetCommitActivityByMonth}
\title{Get an object of commit activity from a specified repository}
\usage{
subsetCommitActivityByMonth(
  commitObj,
  month = month(Sys.Date()),
  year = year(Sys.Date())
)
}
\arguments{
\item{commitObj}{Object returned from an `gh` API pull of commit activity.}

\item{month}{month to subset object for in `m` format.}

\item{year}{year to subset object for in `yyyy` format.}
}
\value{
A subsetted commitObj based on the month/year we are gathering information for.
}
\description{
Get an object of commit activity from a specified repository
}
