% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{webPropertyNameToViewId}
\alias{webPropertyNameToViewId}
\title{Given a Google Analytics `webPropertyName` this function will pull the corresponding GoogleAnalytics account information}
\usage{
webPropertyNameToViewId(webPropertyName)
}
\arguments{
\item{webPropertyName}{a Google Analytics `webPropertyName`. You can find out the webproperty name using `ga_account_list()`.}
}
\value{
The view ID that corrosponds with the provided webPropertyName
}
\description{
Given a Google Analytics `webPropertyName` this function will pull the corresponding GoogleAnalytics account information
}
