% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{getCommitObj}
\alias{getCommitObj}
\title{Lists commits for specified repository and dateRange. Set to return the max amount per page (100).}
\usage{
getCommitObj(owner, repo, dateRange, pageNum)
}
\arguments{
\item{owner}{The GitHub repository owner}

\item{repo}{The GitHub repository name}

\item{dateRange}{A vector of two dates in yyyy-mm-dd format. Can be strings or date objects. Order doesn't matter, the max and min will be used.}

\item{pageNum}{Page number of the results to return}
}
\value{

}
\description{
Lists commits for specified repository and dateRange. Set to return the max amount per page (100).
}
