% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{getGoogleAnalytics}
\alias{getGoogleAnalytics}
\title{Pull Google Analytics data from an account.}
\usage{
getGoogleAnalytics(
  accountName,
  dateRange = c((Sys.Date() - 30), Sys.Date()),
  metrics = c("users", "newUsers", "sessionsPerUser", "pageviews"),
  dimensions = NULL,
  auth = FALSE
)
}
\arguments{
\item{accountName}{specify the account you'd like to get data for, vector length of 1}

\item{dateRange}{specify the dates you'd like to pull data for, vector length 2. Defaults to the last 30 days.}

\item{metrics}{specify the metrics you wish to pull. Defaults:}

\item{auth}{specify if you need to authorize your account.}
}
\value{
a dataframe of variables
}
\description{
Uses the `googleAnalyticsR` package to pull specified data from the Google Analytics API
}
