% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGithub.R
\name{pullGithub}
\alias{pullGithub}
\title{API pull from}
\usage{
pullGithub(
  owner = "FredHutch",
  repo = "coop",
  month = month(Sys.Date()),
  year = year(Sys.Date())
)
}
\arguments{
\item{owner}{The repository name of the relevant GitHub repo.}

\item{repo}{The owner of the relevant GitHub repo.}

\item{month}{specify month as `mm` or `m`. Defaults to current month.}

\item{year}{specify year as `yyyy`. Defaults to current year.}
}
\value{
a dataframe of metrics for the month and year specified.
}
\description{
API pull from
}
