% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullGoogleAnalytics.R
\name{pullGoogleAnalytics}
\alias{pullGoogleAnalytics}
\title{API pull from}
\usage{
pullGoogleAnalytics(
  webPropertyName,
  month = month(Sys.Date()),
  year = year(Sys.Date()),
  .metrics = c("users", "newUsers", "sessions", "pageviews"),
  .topPosts = TRUE
)
}
\arguments{
\item{webPropertyName}{A variable from GoogleAnalytics. You can find out the webproperty name using `ga_account_list()`}

\item{month}{specify month as `mm` or `m`. Defaults to current month.}

\item{year}{specify year as `yyyy`. Defaults to current year.}
}
\value{
a dataframe of metrics for the month and year specified.
}
\description{
API pull from
}
