% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textWrangle.R
\name{wranglePostNames}
\alias{wranglePostNames}
\title{Pull out certain variables from post names that follow the convention of yyyy-mm-dd-title.md}
\usage{
wranglePostNames(postNames, date = TRUE, name = TRUE)
}
\arguments{
\item{postNames}{a vector of post names that follow the convention yyyy-mm-dd-title.md}

\item{date}{output date? T/F}

\item{name}{output name? T/F}
}
\value{
a dataframe of variables
}
\description{
Pull out certain variables from post names that follow the convention of yyyy-mm-dd-title.md
}
